requirements.txt
datasets==2.0.0
numpy==1.21.3
pandas==1.3.4
openpyxl==3.0.9
seaborn==0.11.2
scikit-learn==1.0
spacy==3.1.3
torch==1.11.0
torchmetrics==0.7.2
toolz==0.11.2
transformers==4.17.0
pytorch_lightning==1.5.10
tqdm==4.62.3
pyyaml
scikit-learn==1.0
mlflow==1.24.0
